refactor: serve MCP over Streamable HTTP, drop Vercel/Redis/SSE indirection
The deployed Railway service was only running server.js (a static landing
page); the actual MCP server (api/server.ts) was a Vercel serverless handler
that imported a non-existent vercel.json and required Redis, so it never ran.

Rewrite as a single long-running Node/Express server using the MCP SDK's
Streamable HTTP transport (stateless), which needs no Redis on a single
instance. All eight Payload tools are preserved (extracted to src/tools.ts,
reused via registerTools). Run with tsx; Railway build is now `npm install`
and start is the HTTP server. MCP endpoint: POST /mcp, healthcheck: /health.
