+++ .env.example
@@ -1,33 +1,2 @@
-# Redis connection URL (required for MCP server functionality)
-# Use either REDIS_URL or KV_URL (Vercel KV)
-REDIS_URL=redis://username:password@host:port
-# KV_URL=redis://username:password@host:port
-
-# Vercel KV specific configuration
-# KV_REST_API_URL=https://your-kv-rest-api-url
-# KV_REST_API_TOKEN=your-kv-rest-api-token
-# KV_REST_API_READ_ONLY_TOKEN=your-kv-rest-api-read-only-token
-
-# Connection timeout in milliseconds (optional, default: 30000)
-# REDIS_CONNECT_TIMEOUT=30000
-
-# Keep-alive interval in milliseconds (optional, default: 5000)
-# REDIS_KEEP_ALIVE=5000
-
-# Heartbeat interval in milliseconds (optional, default: 30000)
-# REDIS_HEARTBEAT_INTERVAL=30000
-
-# Redis ping interval in milliseconds (optional, default: 1000)
-# REDIS_PING_INTERVAL=1000
-
-# Redis command timeout in milliseconds (optional, default: 5000)
-# REDIS_COMMAND_TIMEOUT=5000
-
-# Persistence check interval in milliseconds (optional, default: 60000)
-# REDIS_PERSISTENCE_INTERVAL=60000
-
-# Maximum reconnect attempts before forcing clean reconnection (optional, default: 5)
-# REDIS_MAX_RECONNECT_ATTEMPTS=5
-
-# Force TLS verification (optional, default: false for rediss:// URLs)
-# REDIS_TLS_VERIFY=false 
+# Port the MCP server listens on (Railway sets this automatically).
+PORT=8080
