+++ Dockerfile
@@ -1,23 +1,17 @@
-# Generated by https://smithery.ai. See: https://smithery.ai/docs/config#dockerfile
-FROM node:18-alpine
+FROM node:20-alpine
 
-# Create app directory
 WORKDIR /app
 
-# Copy package.json and pnpm-lock.yaml
-COPY package.json pnpm-lock.yaml ./
-
-# Install pnpm
-RUN npm install -g pnpm
-
-# Install dependencies
-RUN pnpm install --no-frozen-lockfile
+# Install dependencies first for better layer caching
+COPY package.json ./
+RUN npm install
 
 # Copy the rest of the application
 COPY . .
 
-# Expose the port the app runs on
+# Railway/clients connect over HTTP on this port
+ENV PORT=8080
 EXPOSE 8080
 
-# Command to run the application
-CMD ["node", "server.js"]
+# Run the MCP server (Streamable HTTP) via tsx
+CMD ["npm", "run", "start"]
