+++ railway.json
@@ -2,16 +2,13 @@
   "$schema": "https://railway.app/railway.schema.json",
   "build": {
     "builder": "NIXPACKS",
-    "buildCommand": "npm install && npm run build",
-    "startCommand": "npm run start"
+    "buildCommand": "npm install"
   },
   "deploy": {
+    "startCommand": "npm run start",
     "healthcheckPath": "/health",
     "healthcheckTimeout": 100,
     "restartPolicyType": "ON_FAILURE",
     "restartPolicyMaxRetries": 10
-  },
-  "variables": {
-    "PORT": "8080"
   }
-} 
+}
